soc/intel/tgl: Add PEG devices to PCI constraints
Based on the constraints for CML.

Fixes the following warnings in Linux on system76/oryp8 and
system76/gaze16, which have an NVIDIA GPU on the bridge.

    pcieport 0000:00:01.0: can't derive routing for PCI INT A
    pcieport 0000:00:01.0: can't derive routing for PCI INT B

This, in turn, resolves an IRQ conflict with the PCH HDA device that
would cause a stack trace on every boot and on S3 suspend.

    irq 10: nobody cared (try booting with the "irqpoll" option)
    <snip>
    [<00000000fb84c354>] azx_interrupt [snd_hda_codec]
    Disabling IRQ #10

Change-Id: Ibc968aaa7bf0259879097ff69d2543dcfa2e5e4b
Signed-off-by: Tim Crawford <tcrawford@system76.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/64671
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>