mb/google/brya/var/ghost: Add auto-generated GPIO config from Arbitrage
Arbitrage is an internal tool at Google to work with schematics
programatically.  In particular, it features an "export-coreboot-gpio"
command, which, does it's best to try and make a gpio.c from the
schematics to avoid human errors when translating to C code.

This commit adds a gpio.c generated by running:

  "arb export-coreboot-gpio ghost4adl:P0_2022_06_17"

This GPIO config will require hand modification.  This is done in a
follow-up CL.  (i.e., this CL intentionally leaves the config exactly
how it was generated by Arbitrage so we get a good diff on the changes
we needed to make)

BUG=b:234626939,b:231719130
BRANCH=firmware-brya-14505.B
TEST=emerge-ghost coreboot

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I35a85202768a366357073d3ebc177d0e0da661f7
Reviewed-on: https://review.coreboot.org/c/coreboot/+/65210
Reviewed-by: Caveh Jalali <caveh@chromium.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>