Makefile: Update error if building real-all when NOCOMPILE is set
The real-all target here had never been updated since the original
NOCOMPILE, which only depended on DOTCONFIG.  Since the reasons that
the NOCOMPILE flag can be set is much larger now, the error given no
longer matches the possible issues.
Give the reason for the failure (nocompile is set), some debug info,
and ask the user to file a bug.

We shouldn't really ever run across this, but I just saw it when I was
working on the NOCOMPILE code.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Change-Id: I5b4be3349fb4cf2d3a8a2a7c183b7a205b9e8733
Reviewed-on: https://review.coreboot.org/c/coreboot/+/65319
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Elyes Haouas <ehaouas@noos.fr>